refactor: modify class naming
